Fix restart interval restart interval is 16 bits, not 8

+++ src/jpg.h
@@ -128,7 +128,7 @@ struct Header {
     byte successiveApproximationHigh;
     byte successiveApproximationLow;
 
-    byte restartInterval = 0;
+    uint restartInterval = 0;
 
     ColorComponent colorComponents[3];
